initial proof of concept. Has a simple ability for the endpoint to flap
Still lots to do, there is absolutely no styling (and that's the lowest
priority). Should consider some kind of state for the endpoints. Not
sure if it matters since it's meant to be ephemeral, but we could
marshal into a file potentially.

Should probably make the flap function tunable, right now it's just a
coinflip